re-organize transitions examples
