+++ examples/use-stdin.ts
